Refactor with features slice structure
