+++ React-Ant/react-ant-ts/src/features/Product/services/ProductsService.tsx
@@ -1,4 +1,4 @@
-import http from "../http-common";
+import http from "../../../services/http-common";
 import { Product } from "../types/Product";
 
 function getAll() {
