+++ React-Ant/react-ant-ts/src/features/TestPage/components/TestPage.tsx
@@ -1,0 +1,5 @@
+function TestPage() {
+
+}
+
+export default TestPage;
